changing structure from server.emit to socket.broadcast
For data binding, there will be a problematic loop where changes from the server are broadcasted back, thus making redundant changes, the emitter of the event should not receive his own changes. It is ok for elemtn creation because the client needs the unique id of the new element back from the server anyways, but it will be problematic for realtime changes such as event triggers, making connections, and moving elements positions around.

This commit is an incomplete implementation of a non-echoing data binding.